Create cmake project and 'solve' messages.h generation.
Actually is not solve yet. The way to go now would be to generate messages.h
automatically. However, the file is generated into the build directory and
we need to put this directory into the include directories for the project
or copy the file to the source directory and I dunno how to do that using
cmake polite way.